// sdk/src/contracts/newport/typechain/OptionMarketViewer.ts
export declare namespace OptionMarketViewer {
  export type OptionMarketAddressesStruct = {
    liquidityPool: string
    liquidityToken: string
    greekCache: string
    optionMarket: string
    optionMarketPricer: string
    optionToken: string
    shortCollateral: string
    poolHedger: string
    quoteAsset: string
    baseAsset: string
  }

  export type OptionMarketAddressesStructOutput = [
    string,
    string,
    string,
    string,
    string,
    string,
    string,
    string,
    string,
    string
  ] & {
    liquidityPool: string
    liquidityToken: string
    greekCache: string
    optionMarket: string
    optionMarketPricer: string
    optionToken: string
    shortCollateral: string
    poolHedger: string
    quoteAsset: string
    baseAsset: string
  }
// ...
  export type BoardViewStructOutput = [
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    boolean,
    BigNumber,
    BigNumber,
    BigNumber,
    OptionGreekCache.NetGreeksStructOutput,
    OptionMarketViewer.StrikeViewStructOutput[]
  ] & {
    market: string
    boardId: BigNumber
    expiry: BigNumber
    baseIv: BigNumber
    priceAtExpiry: BigNumber
    isPaused: boolean
    varianceGwavIv: BigNumber
    forceCloseGwavIv: BigNumber
    longScaleFactor: BigNumber
    netGreeks: OptionGreekCache.NetGreeksStructOutput
    strikes: OptionMarketViewer.StrikeViewStructOutput[]
  }
// ...
  export type MarketParametersStructOutput = [
    OptionMarket.OptionMarketParametersStructOutput,
    LiquidityPool.LiquidityPoolParametersStructOutput,
    LiquidityPool.CircuitBreakerParametersStructOutput,
    OptionGreekCache.GreekCacheParametersStructOutput,
    OptionGreekCache.ForceCloseParametersStructOutput,
    OptionGreekCache.MinCollateralParametersStructOutput,
    OptionMarketPricer.PricingParametersStructOutput,
    OptionMarketPricer.TradeLimitParametersStructOutput,
    OptionMarketPricer.VarianceFeeParametersStructOutput,
    OptionToken.PartialCollateralParametersStructOutput
  ] & {
    optionMarketParams: OptionMarket.OptionMarketParametersStructOutput
    lpParams: LiquidityPool.LiquidityPoolParametersStructOutput
    cbParams: LiquidityPool.CircuitBreakerParametersStructOutput
    greekCacheParams: OptionGreekCache.GreekCacheParametersStructOutput
    forceCloseParams: OptionGreekCache.ForceCloseParametersStructOutput
    minCollatParams: OptionGreekCache.MinCollateralParametersStructOutput
    pricingParams: OptionMarketPricer.PricingParametersStructOutput
    tradeLimitParams: OptionMarketPricer.TradeLimitParametersStructOutput
    varianceFeeParams: OptionMarketPricer.VarianceFeeParametersStructOutput
    partialCollatParams: OptionToken.PartialCollateralParametersStructOutput
  }
// ...
  export type MarketViewWithBoardsStructOutput = [
    boolean,
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    LiquidityPool.LiquidityStructOutput,
    OptionMarketViewer.OptionMarketAddressesStructOutput,
    OptionMarketViewer.MarketParametersStructOutput,
    OptionGreekCache.NetGreeksStructOutput,
    OptionMarketViewer.BoardViewStructOutput[]
  ] & {
    isPaused: boolean
    spotPrice: BigNumber
    minSpotPrice: BigNumber
    maxSpotPrice: BigNumber
    quoteSymbol: string
    quoteDecimals: BigNumber
    baseSymbol: string
    baseDecimals: BigNumber
    rateAndCarry: BigNumber
    liquidity: LiquidityPool.LiquidityStructOutput
    marketAddresses: OptionMarketViewer.OptionMarketAddressesStructOutput
    marketParameters: OptionMarketViewer.MarketParametersStructOutput
    globalNetGreeks: OptionGreekCache.NetGreeksStructOutput
    liveBoards: OptionMarketViewer.BoardViewStructOutput[]
  }
// ...
}

export declare namespace OptionGreekCache {
  export type NetGreeksStruct = {
    netDelta: BigNumberish
    netStdVega: BigNumberish
    netOptionValue: BigNumberish
  }

  export type NetGreeksStructOutput = [BigNumber, BigNumber, BigNumber] & {
    netDelta: BigNumber
    netStdVega: BigNumber
    netOptionValue: BigNumber
  }

  export type StrikeGreeksStruct = {
    callDelta: BigNumberish
    putDelta: BigNumberish
    stdVega: BigNumberish
    callPrice: BigNumberish
    putPrice: BigNumberish
  }

  export type StrikeGreeksStructOutput = [BigNumber, BigNumber, BigNumber, BigNumber, BigNumber] & {
    callDelta: BigNumber
    putDelta: BigNumber
    stdVega: BigNumber
    callPrice: BigNumber
    putPrice: BigNumber
  }

  export type GreekCacheParametersStruct = {
    maxStrikesPerBoard: BigNumberish
    acceptableSpotPricePercentMove: BigNumberish
    staleUpdateDuration: BigNumberish
    varianceIvGWAVPeriod: BigNumberish
    varianceSkewGWAVPeriod: BigNumberish
    optionValueIvGWAVPeriod: BigNumberish
    optionValueSkewGWAVPeriod: BigNumberish
    gwavSkewFloor: BigNumberish
    gwavSkewCap: BigNumberish
  }

  export type GreekCacheParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    maxStrikesPerBoard: BigNumber
    acceptableSpotPricePercentMove: BigNumber
    staleUpdateDuration: BigNumber
    varianceIvGWAVPeriod: BigNumber
    varianceSkewGWAVPeriod: BigNumber
    optionValueIvGWAVPeriod: BigNumber
    optionValueSkewGWAVPeriod: BigNumber
    gwavSkewFloor: BigNumber
    gwavSkewCap: BigNumber
  }

  export type ForceCloseParametersStruct = {
    ivGWAVPeriod: BigNumberish
    skewGWAVPeriod: BigNumberish
    shortVolShock: BigNumberish
    shortPostCutoffVolShock: BigNumberish
    longVolShock: BigNumberish
    longPostCutoffVolShock: BigNumberish
    liquidateVolShock: BigNumberish
    liquidatePostCutoffVolShock: BigNumberish
    shortSpotMin: BigNumberish
    liquidateSpotMin: BigNumberish
  }

  export type ForceCloseParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    ivGWAVPeriod: BigNumber
    skewGWAVPeriod: BigNumber
    shortVolShock: BigNumber
    shortPostCutoffVolShock: BigNumber
    longVolShock: BigNumber
    longPostCutoffVolShock: BigNumber
    liquidateVolShock: BigNumber
    liquidatePostCutoffVolShock: BigNumber
    shortSpotMin: BigNumber
    liquidateSpotMin: BigNumber
  }

  export type MinCollateralParametersStruct = {
    minStaticQuoteCollateral: BigNumberish
    minStaticBaseCollateral: BigNumberish
    shockVolA: BigNumberish
    shockVolPointA: BigNumberish
    shockVolB: BigNumberish
    shockVolPointB: BigNumberish
    callSpotPriceShock: BigNumberish
    putSpotPriceShock: BigNumberish
  }

  export type MinCollateralParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    minStaticQuoteCollateral: BigNumber
    minStaticBaseCollateral: BigNumber
    shockVolA: BigNumber
    shockVolPointA: BigNumber
    shockVolB: BigNumber
    shockVolPointB: BigNumber
    callSpotPriceShock: BigNumber
    putSpotPriceShock: BigNumber
  }
}

export declare namespace LiquidityPool {
  export type LiquidityStruct = {
    freeLiquidity: BigNumberish
    burnableLiquidity: BigNumberish
    reservedCollatLiquidity: BigNumberish
    pendingDeltaLiquidity: BigNumberish
    usedDeltaLiquidity: BigNumberish
    NAV: BigNumberish
    longScaleFactor: BigNumberish
  }

  export type LiquidityStructOutput = [BigNumber, BigNumber, BigNumber, BigNumber, BigNumber, BigNumber, BigNumber] & {
    freeLiquidity: BigNumber
    burnableLiquidity: BigNumber
    reservedCollatLiquidity: BigNumber
    pendingDeltaLiquidity: BigNumber
    usedDeltaLiquidity: BigNumber
    NAV: BigNumber
    longScaleFactor: BigNumber
  }

  export type LiquidityPoolParametersStruct = {
    minDepositWithdraw: BigNumberish
    depositDelay: BigNumberish
    withdrawalDelay: BigNumberish
    withdrawalFee: BigNumberish
    guardianMultisig: string
    guardianDelay: BigNumberish
    adjustmentNetScalingFactor: BigNumberish
    callCollatScalingFactor: BigNumberish
    putCollatScalingFactor: BigNumberish
  }

  export type LiquidityPoolParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    minDepositWithdraw: BigNumber
    depositDelay: BigNumber
    withdrawalDelay: BigNumber
    withdrawalFee: BigNumber
    guardianMultisig: string
    guardianDelay: BigNumber
    adjustmentNetScalingFactor: BigNumber
    callCollatScalingFactor: BigNumber
    putCollatScalingFactor: BigNumber
  }

  export type CircuitBreakerParametersStruct = {
    liquidityCBThreshold: BigNumberish
    liquidityCBTimeout: BigNumberish
    ivVarianceCBThreshold: BigNumberish
    skewVarianceCBThreshold: BigNumberish
    ivVarianceCBTimeout: BigNumberish
    skewVarianceCBTimeout: BigNumberish
    boardSettlementCBTimeout: BigNumberish
    contractAdjustmentCBTimeout: BigNumberish
  }

  export type CircuitBreakerParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    liquidityCBThreshold: BigNumber
    liquidityCBTimeout: BigNumber
    ivVarianceCBThreshold: BigNumber
    skewVarianceCBThreshold: BigNumber
    ivVarianceCBTimeout: BigNumber
    skewVarianceCBTimeout: BigNumber
    boardSettlementCBTimeout: BigNumber
    contractAdjustmentCBTimeout: BigNumber
  }
}